Rakefile

+++ Rakefile
@@ -13,7 +13,7 @@ task :install do
   skip_all = false
   overwrite_all = false
   backup_all = true
-  hostname = `hostname`
+  hostname = `hostname`.chomp
   linkables.each do |linkable|
     overwrite = false
     backup = false
@@ -34,7 +34,9 @@ task :install do
         end
       end
       FileUtils.rm_rf(target) if overwrite || overwrite_all
-      `mv "$HOME/.#{file}" "$HOME/.#{file}.backup"` if backup || backup_all
+      if !File.exists?("$HOME/.#{file}.backup")
+       `mv "$HOME/.#{file}" "$HOME/.#{file}.backup"` if backup || backup_all
+      end
     end
     `ln -s "$PWD/#{linkable}" "#{target}"`
   end
@@ -51,10 +53,8 @@ task :install do
     `ln -s "#{ENV["HOME"]}/usb" "/media/usb0"`
   end
 #create link to configuration of i3 based on hostname 
-  conf_name = "config-#{hostname}"
-  puts "#{conf_name}"
-  if !File.exists?("i3/i3.symlink/#{conf_name}")
-  `ln -s "#{conf_name}" "i3/i3.symlink/config"`
+  if File.exists?("i3/i3.symlink/config-#{hostname}")
+  `ln -s "config-#{hostname}" "i3/i3.symlink/config"`
   end
 end
